Refactor stepper modal components
