+++ packages/ui/src/common/components/Stepper.stories.tsx
@@ -2,7 +2,8 @@ import { Meta, Story } from '@storybook/react'
 import React from 'react'
 
 import { Modal, ModalHeader } from './Modal'
-import { Stepper, StepperModalBody, StepperProps } from './Stepper'
+import { Stepper, StepperProps } from './Stepper'
+import { StepDescriptionColumn, StepperBody, StepperModalBody, StepperModalWrapper } from './StepperModal'
 import { TemplateBlock } from './storybookParts/previewStyles'
 
 export default {
@@ -21,7 +22,11 @@ const Template: Story<StepperProps> = (args) => {
       <Modal onClose={() => undefined} modalSize="l">
         <ModalHeader onClick={() => undefined} title="Modal with stepper" />
         <StepperModalBody>
-          <Stepper {...args} />
+          <StepperModalWrapper>
+            <Stepper {...args} />
+            <StepDescriptionColumn>Col 2</StepDescriptionColumn>
+            <StepperBody>Col 3</StepperBody>
+          </StepperModalWrapper>
         </StepperModalBody>
       </Modal>
     </TemplateBlock>
